+++ src/lib/bitblast/aig/aig_printer.h
@@ -11,17 +11,31 @@
 #ifndef BZLA__BITBLAST_AIG_AIG_PRINTER_H
 #define BZLA__BITBLAST_AIG_AIG_PRINTER_H
 
-#include "bitblast/aig/aig_manager.h"
-
+#include <string>
 #include <vector>
+
+#include "bitblast/aig/aig_node.h"
 
 namespace bzla::bitblast::aig {
 
-class Smt2Printer
+class AigPrinter
 {
  public:
-  static void print(std::stringstream& ss, const AigNode& n);
-  static void print(std::stringstream& ss, const std::vector<AigNode>& bits);
+  /** Add output to print. */
+  void add_output(const AigNode& output);
+
+  /** Assign symbol to AIG node `n`. */
+  void symbol(const AigNode& n, const std::string& symbol);
+
+  /** Write AIG to file in AIGER format. */
+  void write_aiger(const std::string& filename);
+
+  /** Write AIG to file in DIMACS format*/
+  void write_cnf(const std::string& filename);
+
+ private:
+  std::vector<AigNode> d_outputs;
+  std::unordered_map<AigNode, std::string> d_symbols;
 };
 
 }  // namespace bzla::bitblast::aig